requirements.txt
requests
beautifulsoup4
pymongo
python-dotenv
python-telegram-bot